MP2-1203 make all cofinancing numbers as decimals

+++ jems-api/src/main/kotlin/io/cloudflight/jems/api/project/dto/partner/cofinancing/ProjectPartnerCoFinancingOutputDTO.kt
@@ -1,9 +1,10 @@
 package io.cloudflight.jems.api.project.dto.partner.cofinancing
 
 import io.cloudflight.jems.api.programme.dto.fund.ProgrammeFundDTO
+import java.math.BigDecimal
 
 data class ProjectPartnerCoFinancingOutputDTO(
     val fundType: ProjectPartnerCoFinancingFundType,
-    val percentage: Int,
+    val percentage: BigDecimal,
     val fund: ProgrammeFundDTO?
 )
