+++ jems-server/src/main/kotlin/io/cloudflight/jems/server/project/service/partner/cofinancing/model/ProjectPartnerCoFinancing.kt
@@ -2,11 +2,12 @@ package io.cloudflight.jems.server.project.service.partner.cofinancing.model
 
 import io.cloudflight.jems.api.project.dto.partner.cofinancing.ProjectPartnerCoFinancingFundType
 import io.cloudflight.jems.server.programme.service.fund.model.ProgrammeFund
+import java.math.BigDecimal
 
 data class ProjectPartnerCoFinancing(
 
     val fundType: ProjectPartnerCoFinancingFundType,
     val fund: ProgrammeFund? = null,
-    val percentage: Int
+    val percentage: BigDecimal
 
 )
